% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{bonacich_power}
\alias{bonacich_power}
\title{Bonacich power centrality}
\usage{
bonacich_power(
  A,
  beta = 0,
  digraph = TRUE,
  weighted = FALSE,
  scale = c("none", "ssq")
)
}
\arguments{
\item{A}{A square matrix}

\item{beta}{The weight given to the centrality of the neighbours. It should be smaller than the inverse of the leading eigenvalue}

\item{digraph}{Whether the matrix is directed or undirected}

\item{weighted}{Whether the matrix is weighted}

\item{scale}{Whether the scores are returned as they are (\code{none}, default) or scaled so that the sum of their squares is the number of nodes (\code{ssq}), as in other packages}
}
\value{
This function returns the power centrality of the nodes.
}
\description{
Power centrality of Bonacich (1987), where being connected to well-connected others can
increase or decrease the centrality of a node.
}
\details{
The centrality is \eqn{x = (I - \beta A)^{-1} A 1}. When \code{beta} is positive, a node is
central when it is connected to central nodes, as in the eigenvector centrality. When
\code{beta} is negative, being connected to well-connected others reduces the centrality of a
node, which describes bargaining situations. With \code{beta = 0} the measure is the degree.
}
\examples{
A <- matrix(c(
  0, 1, 1, 1, 0,
  1, 0, 0, 0, 0,
  1, 0, 0, 0, 1,
  1, 0, 0, 0, 0,
  0, 0, 1, 0, 0
), byrow = TRUE, ncol = 5)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- rownames(A)

bonacich_power(A, beta = 0.1, digraph = FALSE)
bonacich_power(A, beta = -0.1, digraph = FALSE)
}
\references{
Bonacich, P. (1987). Power and centrality: A family of measures. American Journal of Sociology, 92(5), 1170–1182. \doi{10.1086/228631}
}
\author{
Alejandro Espinosa-Rada
}
