% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_holes.R
\name{burt}
\alias{burt}
\title{Redundancy Measures}
\usage{
burt(A, ego = NULL, digraph = FALSE, weighted = FALSE)
}
\arguments{
\item{A}{A symmetric matrix object}

\item{ego}{Name of ego in the matrix}

\item{digraph}{Whether the  matrix is directed or undirected}

\item{weighted}{Whether the matrix is weighted or not}
}
\value{
This function returns redundancy, effective size and efficincy measures (Burt, 1992).
}
\description{
Redundancy measures of the structural holes theory for binary matrixes
}
\examples{

A <- matrix(c(0,1,0,0,1,1,1,
              1,0,0,1,0,0,1,
              0,0,0,0,0,0,1,
              0,1,0,0,0,0,1,
              1,0,0,0,0,0,1,
              1,0,0,0,0,0,1,
              1,1,1,1,1,1,0), ncol = 7, byrow = TRUE)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- letters[1:ncol(A)]
burt(A, ego = "g")

}
\references{
Burt, R.S., 1992. Structural Holes: the Social Structure of Competition. Harvard University Press, Cambridge.

Borgatti, S., 1997. Unpacking Burt's redundancy measure. Connections, 20(1): 35-38. doi: \url{http://www.analytictech.com/connections/v20(1)/holes.htm}
}
\author{
Alejandro Espinosa-Rada
}
