% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{citation_norm}
\alias{citation_norm}
\title{Citation networks}
\usage{
citation_norm(A1, A2, citation = "citation")
}
\arguments{
\item{A1}{From incident matrix of paper and author.}

\item{A2}{To incident matrix of author to paper.}

\item{citation}{Character string, \dQuote{citation}, \dQuote{cocitation} and \dQuote{bcoupling}}
}
\value{
Return a type of citation network.
}
\description{
Matrix transformation from incident matrices to a citation, fractional counting for co-citation or fractional counting for bibliographic coupling.
}
\examples{

A1 <- matrix(c(1,0,0,0,
               0,1,0,0,
               0,1,1,1,
               0,0,0,0,
               0,0,0,1), byrow = TRUE, ncol=4)

A2 <- matrix(c(1,1,1,0,0,
               0,0,1,0,0,
               0,0,1,1,0,
               0,0,0,1,1), byrow=TRUE, ncol=5)

citation_norm(A1, A2)

}
\references{
Batagelj, V., & Cerinšek, M. (2013). On bibliographic networks. Scientometrics, 96(3), 845–864.
}
\author{
Alejandro Espinosa-Rada
}
