% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{clique_table}
\alias{clique_table}
\title{Clique table}
\usage{
clique_table(A, list_cliques = FALSE, number = FALSE)
}
\arguments{
\item{A}{A symmetric matrix object.}

\item{list_cliques}{Whether to return the list of cliques.}

\item{number}{Number of triangles}
}
\value{
This function return an edge list of actors participating in 3-cliques.

If \code{list_cliques = TRUE} it also  return the list of cliques per nodes.
If \code{number = TRUE} the output returns the number of 3-cliques in the matrix.
}
\description{
Exploration of a 3-cliques, as the maximum number of three or more actors who 
have all possible ties present among themselves
}
\references{
Luce, R.D. and Perry, A.D. (1949). A method of matrix analysis of group structure. Psychometrika, 14: 95-116.

Roethlisberger, F.J. and Dickson, W.J. (1939). Management and the Worker. Harvard University Press, Cambridge, MA.
   
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada

A <- matrix(c(0,1,1,0,0,0,0,1,0,
              1,0,1,0,0,0,0,0,0,
              1,1,0,0,0,0,0,0,0,
              0,0,0,0,1,1,0,0,0,
              0,0,0,1,0,0,0,0,0,
              0,0,0,1,0,0,1,1,0,
              0,0,0,0,0,1,0,1,0,
              1,0,0,0,0,1,1,0,0,
              0,0,0,0,0,0,0,0,0), byrow = TRUE, ncol = 9)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- rownames(A)
clique_table(A, list_cliques = TRUE, number = TRUE)
}
