% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{cliques_table}
\alias{cliques_table}
\title{Find cliques}
\usage{
cliques_table(A, adjacency_list = FALSE, min = NULL, max = NULL)
}
\arguments{
\item{A}{A symmetric matrix object.}

\item{adjacency_list}{Whether to return the adjacency list of cliques per node.}

\item{min}{Numeric constant, lower limit on the size of the cliques to find. NULL means no limit, ie. it is the same as 0.}

\item{max}{Numeric constant, upper limit on the size of the cliques to find. NULL means no limit.}
}
\value{
This function return the list of cliques that each node belong. 

If \code{adjacency_list = TRUE} it also  return the adjacency list of 
cliques per node.
}
\description{
Find cliques
}
\examples{
A <- matrix(c(0,1,1,1,0,
            1,0,1,0,0,
            1,1,0,0,0,
            1,0,0,0,1,
            0,0,0,1,0), byrow = TRUE, ncol = 5)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- letters[1:ncol(A)]

cliques_table(A, adjacency_list = TRUE, min = 3)

}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
