% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{co_ocurrence}
\alias{co_ocurrence}
\title{Co‐occurrence}
\usage{
co_ocurrence(
  A,
  similarity = c("ochiai", "cosine"),
  occurrence = TRUE,
  projection = FALSE
)
}
\arguments{
\item{A}{A matrix}

\item{similarity}{The similarities available are either \code{Ochiai} (default) or \code{cosine}.}

\item{occurrence}{Whether to treat the matrix as a two-mode structure (a.k.a. rectangular matrix, occurrence matrix, affiliation matrix, bipartite network)}

\item{projection}{Whether to apply a projection (inner product multiplication) to the matrix}
}
\value{
This function returns the normalisation of a matrix into a symmetrical co‐occurrence matrix
}
\description{
Co‐occurrence matrix based on overlap function
}
\examples{

A <- matrix(c(
  2, 0, 2,
  1, 1, 0,
  0, 3, 3,
  0, 2, 2,
  0, 0, 1
),
nrow = 5, byrow = TRUE
)

co_ocurrence(A)
}
\references{
Borgatti, S. P., Halgin, D. S., 2011. Analyzing affiliation networks. In: J. Scott and P. J. Carrington (Eds.) The Sage handbook of social network analysis (pp. 417-433), Sage.

Zhou, Q., & Leydesdorff, L. (2016). The normalization of occurrence and Co-occurrence matrices in bibliometrics using Cosine similarities and Ochiai coefficients. Journal of the Association for Information Science and Technology, 67(11), 2805–2814. \url{https://doi.org/10.1002/asi.23603}
}
\author{
Alejandro Espinosa-Rada
}
