% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cumulativeSumMatrices}
\alias{cumulativeSumMatrices}
\title{Cumulative sum of matrices}
\usage{
cumulativeSumMatrices(matrixList)
}
\arguments{
\item{matrixList}{A list of matrices}
}
\value{
This function returns the cumulative sum of matrices
}
\description{
Cumulative sum of matrices
}
\examples{

A <- matrix(c(
  0, 1, 1,
  0, 0, 0,
  0, 1, 0
), byrow = TRUE, ncol = 3)
B <- matrix(c(
  0, 0, 1,
  0, 0, 0,
  0, 0, 0
), byrow = TRUE, ncol = 3)
C <- matrix(c(
  0, 0, 0,
  1, 0, 0,
  0, 0, 0
), byrow = TRUE, ncol = 3)
matrixList <- list(A, B, C)
cumulativeSumMatrices(matrixList)

}
\author{
Alejandro Espinosa-Rada
}
