% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic.R
\name{dist_geographic}
\alias{dist_geographic}
\title{Geographical distances}
\source{
Adapted from Mario Pineda-Krch (Great-circle distance calculations in R)
}
\usage{
dist_geographic(
  latitude,
  longitud,
  method = c("spherical", "harvesine", "manhattan", "minkowski"),
  places = NULL,
  dd_to_radians = FALSE,
  p = NULL
)
}
\arguments{
\item{latitude}{A vector with latitude}

\item{longitud}{A vector with longitud}

\item{method}{Whether to use the Spherical Law of Cosines \code{spherical} (default), Haversine formula \code{harvesine}, Manhattan Distance \code{manhattan} or Minkoowski distance \code{minkowski}}

\item{places}{A vector with the names of the places}

\item{dd_to_radians}{Whether to transform degree decimal format to radians}

\item{p}{Parameter p for the estimation of Minkowski distance (default = 2, which is equivalent to an Euclidian Distance)}
}
\value{
This function return a distance matrix.
}
\description{
This function calculate some geographical distances considering a list of places specifying their latitud and longitud. The function currently works for degree decimal or radians formats.
}
\examples{

set.seed(1234)
x <- cbind(latitud = rnorm(5, -90), longitud = rnorm(5, 45))
dist_geographic(x[, 1], x[, 2], method = "harvesine")
}
