% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{distances}
\alias{distances}
\alias{bfs}
\alias{local_distances}
\alias{all_distances}
\title{Distances}
\usage{
bfs(A, from = NULL)

local_distances(A, select = c("all", "in", "out"), from, to, path = c())

all_distances(A, select = c("all", "in", "out"))
}
\arguments{
\item{A}{A symmetric matrix object}

\item{from}{Node in which the path start}

\item{select}{Whether to consider all sender and receiver ties of ego (\code{all}), only incoming ties (\code{in}), or outgoing ties (\code{out}). By default, \code{all}.}

\item{to}{Node in which the path end}

\item{path}{Path of the nodes}
}
\value{
This function returns the distances o shortest path distance between two nodes
}
\description{
Distances between nodes using breadth-first search (BFS) or Dijkstra's algorithm
}
\examples{
A <- matrix(c(0,1,1,0,0,0,
0,0,0,1,1,0,
0,0,0,0,1,0,
0,0,0,0,0,0,
0,0,0,0,0,1,
0,0,0,0,0,0), byrow = TRUE, nrow = 6)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- letters[1:ncol(A)]

bfs(A, from = 'a')
A <- matrix(c(
  0, 3, 3, 10, 15, 0, 0, 0,
  1, 0, 5, 2, 7, 0, 0, 0,
  3, 5, 0, 0, 0, 0, 0, 0,
  10, 2, 0, 0, 2, 7, 12, 0,
  11, 3, 0, 3, 0, 11, 2, 0,
  0, 0, 0, 7, 11, 0, 3, 2,
  0, 0, 0, 12, 2, 3, 0, 2,
  0, 0, 0, 0, 0, 2, 2, 0
),
byrow = TRUE, ncol = 8, nrow = 8
)
rownames(A) <- c("a", "b", "s", "c", "d", "e", "f", "z")
colnames(A) <- rownames(A)
local_distances(A, from = "a", to = "d")
A <- matrix(c(
  0, 3, 3, 10, 15, 0, 0, 0,
  1, 0, 5, 2, 7, 0, 0, 0,
  3, 5, 0, 0, 0, 0, 0, 0,
  10, 2, 0, 0, 2, 7, 12, 0,
  11, 3, 0, 3, 0, 11, 2, 0,
  0, 0, 0, 7, 11, 0, 3, 2,
  0, 0, 0, 12, 2, 3, 0, 2,
  0, 0, 0, 0, 0, 2, 2, 0
),
byrow = TRUE, ncol = 8, nrow = 8
)
rownames(A) <- c("a", "b", "s", "c", "d", "e", "f", "z")
colnames(A) <- rownames(A)
all_distances(A, select = "in")
}
\references{
Dijkstra, E. W. (1959). A note on two problems in connexion with graphs. Numerische Mathematik. 1: 269–271.
}
\author{
Alejandro Espinosa-Rada
}
