% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{dyadic_census}
\alias{dyadic_census}
\title{Dyad census}
\usage{
dyadic_census(G, directed = TRUE)
}
\arguments{
\item{G}{A symmetric matrix object.}

\item{directed}{Wheter the matrix is directed or not}
}
\value{
This function return the counts of the dyadic census.
}
\description{
Dyad census
}
\examples{

A <- matrix(c(0,1,0,1,0,0,1,0,0,1,0,0,0,1,0,0,0,0,0,0,0,
             1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,
             0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,1,0,0,0,0,
             1,1,0,0,0,0,1,0,0,1,0,0,0,1,1,0,0,0,0,0,0,
             0,1,0,0,0,0,0,0,1,0,0,1,0,0,1,0,1,0,1,1,0,
             0,1,0,0,0,0,0,0,0,1,0,0,0,0,1,0,0,0,1,1,1,
             0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
             0,0,1,0,1,0,1,0,0,1,0,0,0,1,0,0,0,1,0,1,0,
             1,1,1,1,1,0,1,0,0,1,1,0,1,0,1,1,1,0,0,1,0,
             1,0,0,1,0,0,0,0,0,0,0,0,0,0,1,0,0,0,1,0,0,
             0,0,0,0,1,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,
             0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,1,
             1,0,1,0,1,1,0,0,1,0,0,1,0,0,0,0,1,0,0,1,0,
             1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
             1,1,1,1,1,1,1,1,1,1,0,1,1,1,0,0,1,1,1,1,1,
             0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
             1,1,1,0,1,0,0,0,1,1,0,1,1,0,0,0,0,1,0,0,0,
             0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,1,0,0,0,0,0,
             0,1,0,0,0,0,0,0,0,1,0,0,0,0,1,1,0,0,0,0,0,
             0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
             0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0), 
             ncol=21, byrow=TRUE)

dyadic_census(A)

}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
