% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_component}
\alias{extract_component}
\title{Extract components}
\usage{
extract_component(A, maximum = TRUE, position = NULL)
}
\arguments{
\item{A}{A matrix}

\item{maximum}{Whether to extract the maximum component}

\item{position}{The position of the size of the component, from the largest (1). Used when \code{maximum = FALSE}}
}
\value{
The matrix of the component, or a list with the matrices of the components when several have the same size
}
\description{
This function extract the matrix of different components
}
\examples{

A <- FIFAex$Matrix
rownames(A) <- FIFAex$label
colnames(A) <- rownames(A)
extract_component(A, maximum = TRUE)
extract_component(A, maximum = FALSE, position = 2)
}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
