% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifa_multilevel.R
\docType{data}
\name{FIFAego}
\alias{FIFAego}
\title{Ego FIFA}
\format{
A list of a 48 x 48 symmetric matrix of the ego network of FIFA as an entity, and a string vector with the label of the actors
}
\source{
Espinosa, Alejandro & Ortiz, Francisca (2016). "Jurisdictional autonomy in the regulatory transnational regime of FIFA". REDES- Revista Hispana para el Análisis de Redes Sociales, 27(1), 100- 112. (Original title in Spanish: "Autonomía jurisdiccional en el régimen regulatorio transnacional de la FIFA") <doi: https://doi.org/10.5565/rev/redes.595>
}
\usage{
FIFAego
}
\description{
Multilevel Network of the regulatory transnational regime of the International Federation of Association Football (FIFA)
}
\keyword{datasets}
