% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifa_multilevel.R
\docType{data}
\name{FIFAin}
\alias{FIFAin}
\title{Inside FIFA}
\format{
A list of a 41 x 41 symmetric matrix of roles and organizations inside FIFA, and a string vector with the label of the actors
}
\source{
Espinosa, Alejandro & Ortiz, Francisca (2016). "Jurisdictional autonomy in the regulatory transnational regime of FIFA". REDES- Revista Hispana para el Analisis de Redes Sociales, 27(1), 100- 112. (Original title in Spanish: "Autonomia jurisdiccional en el reegimen regulatorio transnacional de la FIFAFA") <doi: https://doi.org/10.5565/rev/redes.595>
}
\usage{
FIFAin
}
\value{
Roles and organizations inside FIFA, and a string vector with the label of the actors
}
\description{
Multilevel Network of the regulatory transnational regime of the International Federation of Association Football (FIFA)
}
\keyword{datasets}
