% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{gen_degree}
\alias{gen_degree}
\title{Degree centrality for one-mode and bipartite networks.}
\usage{
gen_degree(
  A,
  weighted = FALSE,
  type = "out",
  normalized = FALSE,
  loops = TRUE,
  digraph = TRUE,
  alpha = 0.5,
  bipartite = FALSE
)
}
\arguments{
\item{A}{A symmetric matrix object}

\item{weighted}{Wheter the matrix is weighted or not}

\item{type}{Character string, \dQuote{out} (outdegree), \dQuote{in} (indegree) and \dQuote{all} (degree)}

\item{normalized}{Wheter normalize the measure for the one-mode network (Freeman, 1978) or a bipartite network (Borgatti and Everett, 1997)}

\item{loops}{Wheter if the diagonal of the matrix is considered or not}

\item{digraph}{Wheter the  matrix is directed or undirected}

\item{alpha}{Sets the alpha parameter in the generalised measures from Opsahl et al. (2010)}

\item{bipartite}{Wheter the matrix is bipartite or not.}
}
\value{
This function returns term 1, 2 and 3, the normalization and the maximum value of the specification of Everett and Borgatti (2020),
and the constraint of Burt (1992).
}
\description{
Degree centrality for one-mode and bipartite networks.
}
\examples{

A <- matrix(c(0,1,0,0,0,1,
             1,0,1,0,0,1,
             1,1,0,0,0,1,
             1,1,0,0,1,1,
             0,0,0,1,0,1,
             1,1,1,1,1,0), ncol=6, byrow=TRUE)

gen_degree(A)

}
\references{
Borgatti, S. P., and Everett, M. G. (1997). Network analysis of 2-mode data. Social Networks, 19(3), 243–269.

Freeman, L. C. (1978). Centrality in social networks conceptual clarification. Social Networks, 1(3), 215–239.

Opsahl, T., Agneessens, F., and Skvoretz, J. (2010). Node centrality in weighted networks: Generalizing degree and shortest paths. Social Networks, 32(3), 245–251.
}
