% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{geodesic}
\alias{geodesic}
\alias{local_geodesic}
\alias{all_geodesic}
\title{Geodesic}
\usage{
local_geodesic(A, select = c("all", "in", "out"), from, to, path = c())

all_geodesic(A, select = c("all", "in", "out"))
}
\arguments{
\item{A}{A symmetric matrix object}

\item{select}{Whether to consider all sender and receiver ties of ego (\code{all}), only incoming ties (\code{in}), or outgoing ties (\code{out}). By default, \code{all}.}

\item{from}{Node in which the path start}

\item{to}{Node in which the path end}

\item{path}{Path of the nodes}
}
\value{
This function returns the geodesic o shortest path distance between two nodes
}
\description{
Shortest path between two nodes using Dijkstra's algorithm
}
\examples{
A <- matrix(c(
  0, 3, 3, 10, 15, 0, 0, 0,
  1, 0, 5, 2, 7, 0, 0, 0,
  3, 5, 0, 0, 0, 0, 0, 0,
  10, 2, 0, 0, 2, 7, 12, 0,
  11, 3, 0, 3, 0, 11, 2, 0,
  0, 0, 0, 7, 11, 0, 3, 2,
  0, 0, 0, 12, 2, 3, 0, 2,
  0, 0, 0, 0, 0, 2, 2, 0
),
byrow = TRUE, ncol = 8, nrow = 8
)
rownames(A) <- c("a", "b", "s", "c", "d", "e", "f", "z")
colnames(A) <- rownames(A)
local_geodesic(A, from = "a", to = "d")
A <- matrix(c(
  0, 3, 3, 10, 15, 0, 0, 0,
  1, 0, 5, 2, 7, 0, 0, 0,
  3, 5, 0, 0, 0, 0, 0, 0,
  10, 2, 0, 0, 2, 7, 12, 0,
  11, 3, 0, 3, 0, 11, 2, 0,
  0, 0, 0, 7, 11, 0, 3, 2,
  0, 0, 0, 12, 2, 3, 0, 2,
  0, 0, 0, 0, 0, 2, 2, 0
),
byrow = TRUE, ncol = 8, nrow = 8
)
rownames(A) <- c("a", "b", "s", "c", "d", "e", "f", "z")
colnames(A) <- rownames(A)
all_geodesic(A, select = "in")
}
\references{
Dijkstra, E. W. (1959). A note on two problems in connexion with graphs. Numerische Mathematik. 1: 269–271.
}
\author{
Alejandro Espinosa-Rada
}
