% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.R
\name{indirect_rel}
\alias{indirect_rel}
\title{Indirect relations}
\usage{
indirect_rel(
  A,
  type = c("distance", "adjacency", "walks", "shared"),
  digraph = TRUE,
  alpha = 0.1
)
}
\arguments{
\item{A}{A square matrix}

\item{type}{The relation: \code{adjacency}, \code{distance} (default), \code{walks} or \code{shared}}

\item{digraph}{Whether the matrix is directed or undirected}

\item{alpha}{The discount of the longer walks, for the \code{walks} relation}
}
\value{
This function returns a square matrix of relations.
}
\description{
Relations between the nodes that are derived from the ties of the network, to be compared
with \code{pos_dominance()}.
}
\details{
\code{adjacency}: the ties themselves.

\code{distance}: the length of the shortest path between the nodes.

\code{walks}: the number of walks of any length that join two nodes, where a walk of length
\eqn{k} is discounted by \eqn{\alpha^k}, as in the Katz centrality.

\code{shared}: the number of neighbours that two nodes have in common.
}
\examples{
A <- matrix(c(
  0, 1, 1, 1, 0,
  1, 0, 0, 0, 0,
  1, 0, 0, 0, 1,
  1, 0, 0, 0, 0,
  0, 0, 1, 0, 0
), byrow = TRUE, ncol = 5)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- rownames(A)

indirect_rel(A, type = "distance", digraph = FALSE)
indirect_rel(A, type = "walks", digraph = FALSE)
}
\references{
Brandes, U. (2016). Network positions. Methodological Innovations, 9, 1–19. \doi{10.1177/2059799116630650}
}
\author{
Alejandro Espinosa-Rada
}
