% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard similarity}
\usage{
jaccard(A, B, directed = TRUE, diag = FALSE)
}
\arguments{
\item{A}{Binary matrix A}

\item{B}{Binary matrix B}

\item{directed}{Whether the matrix is symmetric}

\item{diag}{Whether the diagonal should be considered}
}
\value{
Jaccard similarity, proportion among the ties present at a given observation of ties that are also present in the other matrix, and a table with the tie changes between matrices
}
\description{
Jaccard similarity to identify the tie changes between two matrices.
}
\examples{

A <- matrix(c(0,1,1,0,
              1,0,0,0,
              1,0,0,0,
              0,0,1,0), byrow=TRUE, ncol= 4)
B <- matrix(c(0,1,1,0,
              1,0,0,0,
              1,0,0,0,
              0,0,0,0), byrow=TRUE, ncol= 4)
jaccard(A, B, directed = TRUE)

}
\references{
Batagelj, V., and Bren, M. (1995). Comparing resemblance measures. Journal of Classification 12, 73–90.
}
\author{
Alejandro Espinosa-Rada
}
