% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krackhardt_friends.R
\docType{data}
\name{krackhardt_friends}
\alias{krackhardt_friends}
\title{Krackhardt friends}
\format{
A 21 x 21 directed matrix of the managers
}
\source{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\usage{
data(krackhardt_friends)
}
\description{
Friendship network of the relations measured for Krackhardt's high-tech managers.
}
\keyword{datasets}
