% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazega_law.R
\docType{data}
\name{lazega_lawfirm}
\alias{lazega_lawfirm}
\title{Lazega law firm}
\format{
Three 71 X 71 matrices:
\describe{
    \item{cowork}{A matrix indicating "cowork" relationships among the attorneys, based on their work together on cases and other professional activities.}
    \item{advice}{A matrix indicating "advice" relationships, where the matrix shows to whom attorneys went for professional advice.}
    \item{friends}{A matrix indicating "friends" relationships, showing social connections outside of work.}
    \item{attributes}{A data frame with attributes of the actors, including:}
    \describe{
        \item{seniority}{Seniority.}
        \item{status}{1=partner; 2=associate}
        \item{gender}{1=man; 2=woman}
        \item{office}{1=Boston; 2=Hartford; 3=Providence}
        \item{years}{Years with the firm}
        \item{age}{Age}
        \item{practice}{1=litigation; 2=corporate}
        \item{law_school}{1 = Harvard, Yale; 2 = UConn; 3 = Other}
    }
}
}
\source{
Lazega, Emmanuel (2001) The Collegial Phenomenon: The Social Mechanisms of Cooperation Among Peers in a Corporate Law Partnership. Oxford University Press.
}
\usage{
data(lazega_lawfirm)
}
\description{
The data is part of a study carried out in a Northeastern US corporate law firm,
referred to as SG&R, 1988-1991 in New England.
The data were collected by Emmanuel Lazega (2001).
This is a multiplex network of attorneys (partners and associates) of this firm.
It includes (among others) measurements of networks among the 71 attorneys (partners and associates) of this firm, i.e. their strong-coworker network, advice network, friendship network, and indirect control networks.
}
\keyword{datasets}
