% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{leading_eigen}
\alias{leading_eigen}
\title{Communities with the leading eigenvector}
\usage{
leading_eigen(A, digraph = FALSE, weighted = FALSE, max_groups = NULL)
}
\arguments{
\item{A}{A square matrix}

\item{digraph}{Whether the matrix is directed or undirected}

\item{weighted}{Whether the matrix is weighted}

\item{max_groups}{Maximum number of groups}
}
\value{
This function returns the group of each node, the number of groups and the modularity of the partition.
}
\description{
Groups of nodes that are connected among themselves more often than expected, found with the
leading eigenvector of the modularity matrix (Newman, 2006).
}
\details{
The modularity matrix is \eqn{B = A - kk^T / 2m}, the observed ties minus the ties expected
from the degrees. The sign of its leading eigenvector splits the network into two groups, and
each group is split again while the modularity of the partition increases.
}
\examples{
A <- matrix(c(
  0, 1, 1, 0, 0, 0,
  1, 0, 1, 0, 0, 0,
  1, 1, 0, 1, 0, 0,
  0, 0, 1, 0, 1, 1,
  0, 0, 0, 1, 0, 1,
  0, 0, 0, 1, 1, 0
), byrow = TRUE, ncol = 6)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- rownames(A)

leading_eigen(A)
}
\references{
Newman, M. E. J. (2006). Finding community structure in networks using the eigenvectors of matrices. Physical Review E, 74(3), 036104. \doi{10.1103/PhysRevE.74.036104}
}
\author{
Alejandro Espinosa-Rada
}
