% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_adjlist}
\alias{matrix_adjlist}
\title{Transform a matrix to an adjacency list}
\usage{
matrix_adjlist(A)
}
\arguments{
\item{A}{A matrix}
}
\value{
This function transform a matrix to an adjacency list
}
\description{
Transform a matrix to an adjacency list
}
\examples{
A <- matrix(c(
  0, 1, 1, 0, 0, 0, 0, 1, 0,
  1, 0, 1, 0, 0, 0, 0, 0, 0,
  1, 1, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 1, 1, 0, 0, 0,
  0, 0, 0, 1, 0, 0, 0, 0, 0,
  0, 0, 0, 1, 0, 0, 1, 1, 0,
  0, 0, 0, 0, 0, 1, 0, 1, 0,
  1, 0, 0, 0, 0, 1, 1, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0
), byrow = TRUE, ncol = 9)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- rownames(A)
matrix_adjlist(A)
}
\author{
Alejandro Espinosa-Rada
}
