% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_projection}
\alias{matrix_projection}
\title{Unipartite projections}
\usage{
matrix_projection(A)
}
\arguments{
\item{A}{A matrix object}
}
\value{
This function return a list of matrices of the two projections of the original matrix.
}
\description{
Two-mode networks can be represented (or 'projected') as one-mode networks.
}
\examples{
A <- matrix(c(
  2, 0, 2,
  1, 1, 0,
  0, 3, 3,
  0, 2, 2,
  0, 0, 1
), byrow = TRUE, ncol = 3)
matrix_projection(A)
}
\references{
Davis, Allison; Gardner, Burleigh B. and Mary. R. Gardner (1941). Deep South: A Social Anthropological Study of Caste and Class. The University of Chicago Press, Chicago.

Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
