% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_report}
\alias{matrix_report}
\title{Matrix report}
\usage{
matrix_report(A)
}
\arguments{
\item{A}{A matrix}
}
\value{
This function return a report of some of the characteristics of the matrix.
}
\description{
The primary matrix used in social network analysis are the
adjacency matrix or sociomatrix, and the incidence matrix.
}
\examples{
A <- matrix(c(
  1, 1, 0, 0, -1,
  1, 0, 0, 1, 1,
  0, 0, NA, 1, 1,
  0, 1, 1, 0, 1,
  1, 1, 1, 1, 0
), byrow = TRUE, ncol = 5)

B <- matrix(c(
  1, 0, 0,
  1, 1, 0,
  0, NA, 0,
  0, 1, 0,
  0, 1, 1
), byrow = TRUE, ncol = 3)
matrix_report(A)
matrix_report(B)
}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
