% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_to_edgelist}
\alias{matrix_to_edgelist}
\title{Transform symmetric matrix to an edge-list}
\usage{
matrix_to_edgelist(A, digraph = FALSE, valued = FALSE, loops = FALSE)
}
\arguments{
\item{A}{A symmetric matrix}

\item{digraph}{Whether the matrix is directed or not}

\item{valued}{Add a third columns with the valued of the relationship}

\item{loops}{Whether the loops are retained or not}
}
\value{
This function transform the matrix into an edgelist.
}
\description{
Transform symmetric matrix to an edge-list
}
\examples{
A <- matrix(c(0,2,1,
              1,0,0,
              1,0,1), byrow=TRUE, ncol=3)
matrix_to_edgelist(A, digraph=TRUE, valued = TRUE, loops=TRUE)

}
\author{
Alejandro Espinosa-Rada
}
