% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{mixed_census}
\alias{mixed_census}
\title{Multilevel triadic and quadrilateral census}
\usage{
mixed_census(A1, B1, B2 = NULL, quad = FALSE)
}
\arguments{
\item{A1}{An adjacent matrix object.}

\item{B1}{An incident matrix object.}

\item{B2}{An incident matrix object.}

\item{quad}{Wheter the matrix is a quadrilateral census or not.}
}
\value{
This function return the counts of the dyadic census.
}
\description{
Multilevel triadic and quadrilateral census
}
\examples{

B1 <- matrix(c(1,1,0,
               0,0,1,
               0,0,1,
               1,0,0), byrow=TRUE, ncol=3)
A1 <- matrix(c(0,1,0,1,
               1,0,0,1,
               0,1,0,1,
               1,0,1,0), byrow=TRUE, ncol=4)
B2 <- matrix(c(1,0,0,0,0,
               0,1,0,1,0,
               0,0,0,0,0,
               0,0,0,0,0), byrow=TRUE, ncol=5)

mixed_census(A1, B1, B2, quad=TRUE)
 
}
\references{
Espinosa-Rada, Alejandro; Bellotti, Elisa; Everett, Martin & Stadtfeld, Christoph (forthcoming). "Co-evolution of Scientific Networks: Multilevel Analysis of a National Discipline"

Hollway, J., Lomi, A., Pallotti, F., & Stadtfeld, C. (2017). Multilevel social spaces: The network dynamics of organizational fields. Network Science, 5(2), 187–212. https://doi.org/10.1017/nws.2017.8
}
\author{
James Hollway

Alessandro Lomi

Francesca Pallotti

Christoph Stadtfeld

Alejandro Espinosa-Rada
}
