% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{MixedLayerTriadCensus}
\alias{MixedLayerTriadCensus}
\title{Multiplex triad census for a directed and an undirected network.}
\usage{
MixedLayerTriadCensus(A, B)
}
\arguments{
\item{A}{A directed matrix object.}

\item{B}{An undirected matrix object.}
}
\value{
This function gives the counts of the mixed multiplex triad census for a directed and an undirected network. The contribution of each of the elementary form is 0.5 and in the multiplex triads sum 1.
}
\description{
This function counts the different subgraphs of three nodes in a multiplex directed and undirected network.
}
\examples{

# SOAR
A <- matrix(c(0,1,1,1,1,0,0,1,1,0,1,1,
              0,0,0,0,0,0,0,0,0,0,0,0,
              0,0,0,0,0,0,0,0,0,0,1,0,
              0,0,0,0,0,0,0,0,0,0,0,0,
              0,1,1,1,0,0,0,1,1,0,1,1,
              0,0,0,1,0,0,0,0,1,0,0,0,
              0,0,0,1,0,1,0,0,1,0,0,0,
              0,0,0,0,0,0,0,0,0,0,0,0,
              0,1,1,1,1,0,0,1,0,0,1,1,
              0,1,0,0,0,1,0,0,0,0,0,0,
              0,1,0,0,1,0,0,0,1,0,0,1,
              0,0,0,0,0,0,0,0,0,0,0,0), 
            byrow=TRUE, ncol=12)

B <- matrix(c(0,0,0,0,1,0,0,0,1,0,0,0,
              0,0,0,0,0,0,0,0,1,0,0,1,
              0,0,0,0,0,0,0,0,0,0,0,0,
              0,0,0,0,0,1,0,0,0,0,0,0,
              1,0,0,0,0,0,0,0,1,0,1,0,
              0,0,0,1,0,0,1,0,0,0,0,0,
              0,0,0,0,0,1,0,0,0,0,0,0,
              0,0,0,0,0,0,0,0,0,0,0,0,
              1,1,0,0,1,0,0,0,0,0,0,0,
              0,0,0,0,0,0,0,0,0,0,0,0,
              0,0,0,0,1,0,0,0,0,0,0,0,
              0,1,0,0,0,0,0,0,0,0,0,0), 
            byrow = TRUE, ncol=12)

MixedLayerTriadCensus(A, B)
}
\references{
Espinosa-Rada, Alejandro; Bellotti, Elisa; Everett, Martin & Stadtfeld, Christoph (forthcoming). "Co-evolution of Scientific Networks: Multilevel Analysis of a National Discipline"
}
\author{
Alejandro Espinosa-Rada
}
