% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohesive_subgroups.R
\name{percolation_clique}
\alias{percolation_clique}
\title{Clique percolation}
\usage{
percolation_clique(A)
}
\arguments{
\item{A}{A matrix}
}
\value{
A matrix that assign each node to a clique
}
\description{
Clique Percolation Method (CPM) is an algorithm for finding overlapping communities within networks, introduced by Palla et al. (2005). This function firstly identify cliques of size k, then creates a incidence matrix as an affiliation network.
}
\examples{

A <- matrix(c(
  0, 1, 1, 1, 0, 0, 0, 0, 0,
  1, 0, 1, 0, 0, 0, 0, 0, 0,
  1, 1, 0, 1, 0, 0, 0, 0, 0,
  1, 0, 1, 0, 1, 1, 0, 0, 0,
  0, 0, 0, 1, 0, 1, 1, 1, 0,
  0, 0, 0, 1, 1, 0, 1, 1, 0,
  0, 0, 0, 0, 1, 1, 0, 1, 1,
  0, 0, 0, 0, 1, 1, 1, 0, 0,
  0, 0, 0, 0, 0, 0, 1, 0, 0
),
byrow = TRUE, ncol = 9
)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- letters[1:ncol(A)]
percolation_clique(A)
}
\references{
Palla, G., Derényi, I., Farkas, I., & Vicsek, T. (2005). Uncovering the overlapping community structure of complex networks in nature and society. Nature, 435(7043), 814-818.
}
\author{
Alejandro Espinosa-Rada
}
