% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{perm_label}
\alias{perm_label}
\title{Permute labels of a matrix}
\usage{
perm_label(A, m = 1, unique = FALSE)
}
\arguments{
\item{A}{A matrix}

\item{m}{Number of permutations}

\item{unique}{Whether to return unique cases}
}
\value{
This function returns the permutation of labels.
}
\description{
This function permutes the labels of a matrix.
}
\examples{

W <- matrix(c(
  0, 1, 0, 0, 0,
  0, 0, 1, 0, 0,
  1, 0, 0, 0, 0,
  0, 0, 0, 0, 1,
  0, 0, 0, 1, 0
), byrow = TRUE, ncol = 5)
rownames(W) <- c("P", "Q", "R", "S", "T")
colnames(W) <- rownames(W)
perm_label(W, m = 1000, unique = TRUE)

}
\author{
Alejandro Espinosa-Rada
}
