% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{perm_matrix}
\alias{perm_matrix}
\title{Permutation matrix}
\usage{
perm_matrix(n, m = 1, unique = FALSE)
}
\arguments{
\item{n}{The size of the square matri}

\item{m}{Number of permutations}

\item{unique}{Whether to return unique cases}
}
\value{
This function returns a list of permutation matrices
}
\description{
This function create permutation matrices.
}
\examples{

W <- matrix(c(
  0, 1, 0, 0, 0,
  0, 0, 1, 0, 0,
  1, 0, 0, 0, 0,
  0, 0, 0, 0, 1,
  0, 0, 0, 1, 0
), byrow = TRUE, ncol = 5)
rownames(W) <- c("P", "Q", "R", "S", "T")
colnames(W) <- rownames(W)
perm_matrix(5, m = 1000, unique = TRUE)

}
\author{
Alejandro Espinosa-Rada
}
