% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{pkp}
\alias{pkp}
\title{Reciprocity of Katz and Powell}
\usage{
pkp(G, fixed = FALSE, d = NULL, dichotomic = TRUE)
}
\arguments{
\item{G}{A symmetric matrix object.}

\item{fixed}{Whether the choices are fixed or not}

\item{d}{Numeric value of the number of fixed choices.}

\item{dichotomic}{Whether the matrix is weighted or binary}
}
\value{
This function gives a measurment of the tendency toward reciprocation of choices.
}
\description{
Reciprocity of Katz and Powell
}
\examples{

data(krackhardt_friends)
pkp(krackhardt_friends, fixed=TRUE, d=5)

}
\references{
Katz, L. and Powell, J.H. (1955). "Measurement of the tendency toward reciprocation of choice." Sociometry, 18:659-665.
}
\author{
Alejandro Espinosa-Rada
}
