% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_networks.R
\name{pref_attachment}
\alias{pref_attachment}
\title{Preferential attachment network}
\usage{
pref_attachment(n, m = 1, power = 1, digraph = FALSE)
}
\arguments{
\item{n}{The number of nodes}

\item{m}{The number of ties that every new node creates}

\item{power}{The power of the degree in the probability of being chosen}

\item{digraph}{Whether the ties of the new nodes are directed towards the nodes that are already there}
}
\value{
This function returns a matrix.
}
\description{
Random network of Barabasi and Albert (1999), in which the nodes that already have more ties
are more likely to receive the ties of the nodes that arrive.
}
\details{
The network starts with \code{m} nodes without ties. Every new node creates \code{m} ties
with the nodes that are already there, choosing each of them with a probability proportional
to their degree raised to \code{power}. With \code{power = 0} the nodes are chosen at random,
and the higher the power, the more the ties concentrate in a few nodes.
}
\examples{
set.seed(18051889)
A <- pref_attachment(20, m = 2)
gen_degree(A, digraph = FALSE)
}
\references{
Barabasi, A. L. and Albert, R. (1999). Emergence of scaling in random networks. Science, 286(5439), 509–512. \doi{10.1126/science.286.5439.509}
}
\author{
Alejandro Espinosa-Rada
}
