% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{recip_coef}
\alias{recip_coef}
\title{Reciprocity}
\usage{
recip_coef(
  A,
  diag = NULL,
  method = c("total_ratio", "ratio_nonnull", "global")
)
}
\arguments{
\item{A}{A matrix}

\item{diag}{Whether to consider the diagonal of the matrix}

\item{method}{Whether to use \code{total_ratio}, \code{ratio_nonnull} or \code{global} reciprocity}
}
\value{
Return a reciprocity coefficient
}
\description{
This measure calculated the reciprocity of an asymmetric matrix (directed graph).
}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
A <- matrix(c(0,1,1,0,
1,0,1,0,
0,0,0,0,
1,0,0,0), byrow = TRUE, ncol = 4)
recip_coef(A)
}
