% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.R
\name{set_inclusion}
\alias{set_inclusion}
\title{Set inclusion of neighbourhoods}
\usage{
set_inclusion(N, M = N, proper = FALSE)
}
\arguments{
\item{N}{An incidence matrix of the neighbourhoods to be included}

\item{M}{An incidence matrix, with the same dimensions as \code{N}, of the including neighbourhoods}

\item{proper}{Whether the inclusion should be proper}
}
\value{
This function returns a binary matrix \code{P} where \code{P[u, v] = 1} if \code{N(u)} is included in \code{M(v)}.
}
\description{
Inclusion of the neighbourhoods of the rows of an incidence matrix.
}
\details{
The neighbourhoods of the rows can be defined on any set of columns (e.g. papers of authors, or
events of actors). \code{N} contains the neighbourhood that should be included, and \code{M} the
neighbourhood in which it should be included. \code{M} is the same as \code{N} for open
neighbourhoods, or can add other elements, such as the node itself, for closed neighbourhoods.

With \code{proper = TRUE} the inclusion should be proper, \eqn{N(u) \subsetneq M(v)}, i.e.
\eqn{M(v)} has at least one element that is not in \eqn{N(u)}.
}
\examples{
X <- matrix(c(
  1, 1, 1, 0,
  1, 1, 0, 0,
  1, 0, 0, 0,
  0, 1, 0, 0,
  0, 0, 1, 1
), byrow = TRUE, ncol = 4)
rownames(X) <- c("a1", "a2", "a3", "a4", "a5")
colnames(X) <- c("w1", "w2", "w3", "w4")

set_inclusion(X)
}
\author{
Alejandro Espinosa-Rada
}
