% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohesive_subgroups.R
\name{shared_partners}
\alias{shared_partners}
\title{Shared partners}
\usage{
shared_partners(
  A,
  loops = FALSE,
  directed = TRUE,
  type = c("dsp", "esp", "nsp")
)
}
\arguments{
\item{A}{A binary matrix}

\item{loops}{Whether to consider the loops}

\item{directed}{Whether the matrix is directed}

\item{type}{Whether to return the \code{dyad-wise (dsp)} (default), \code{edge-wise (esp)} or \code{non-edgewise (nsp)} shared partners (Hunter and Handcock, 2006)}
}
\value{
This function return the distribution of shared partners.
}
\description{
Shared partners
}
\examples{
A <- matrix(c(
  0, 1, 0, 0, 0, 0,
  1, 0, 1, 1, 0, 1,
  0, 1, 0, 1, 0, 0,
  0, 1, 1, 0, 1, 1,
  0, 0, 0, 1, 0, 1,
  0, 1, 0, 1, 1, 0
), byrow = TRUE, ncol = 6)
shared_partners(A, type = "dsp")
shared_partners(A, type = "esp")
shared_partners(A, type = "nsp")
}
\references{
Hunter, D. R. and M. S. Handcock (2006), Inference in curved exponential family models for networks, Journal of Computational and Graphical Statistics, 15: 565– 583.
}
\author{
Alejandro Espinosa-Rada
}
