% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{triad201_table}
\alias{triad201_table}
\title{Find triads 201}
\usage{
triad201_table(A, adjacency_list = FALSE, min = NULL, max = NULL)
}
\arguments{
\item{A}{A symmetric matrix object.}

\item{adjacency_list}{Whether to return the adjacency list of triads 201 per node.}

\item{min}{Numeric constant, lower limit on the size of the triads 201 to find. NULL means no limit, ie. it is the same as 0.}

\item{max}{Numeric constant, upper limit on the size of the triads 201 to find. NULL means no limit.}
}
\value{
This function return the list of triads 201 that each node belong. 

If \code{adjacency_list = TRUE} it also  return the adjacency list of 
triads 201 per node.
}
\description{
The triad 201 = A<->B<->C is one of the possible triads according to the MAN census of Davis and Leinhardt (1972).
}
\examples{
A <- matrix(c(0,1,1,1,0,
            1,0,1,0,0,
            1,1,0,0,0,
            1,0,0,0,1,
            0,0,0,1,0), byrow = TRUE, ncol = 5)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- letters[1:ncol(A)]

triad201_table(A, adjacency_list = TRUE, min = 3)

}
\references{
Davis, J.A. and Leinhardt, S. (1972). The Structure of Positive Interpersonal Relations in Small Groups. In J. Berger (Ed.), Sociological Theories in Progress, Volume 2, 218-251. Boston: Houghton Mifflin.

Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
