% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{ZBarc}
\alias{ZBarc}
\title{Z test for Generalized Bernoulli Models}
\usage{
ZBarc(G, p = 0.5)
}
\arguments{
\item{G}{A symmetric matrix object.}

\item{p}{p-value}
}
\value{
This function gives a Z test and p-value for arcs in uniform distribution
}
\description{
Z test for Generalized Bernoulli Models
}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge university press.
}
\author{
Alejandro Espinosa-Rada
}
