% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{zIntervals}
\alias{zIntervals}
\title{Z test for Generalized Bernoulli Models}
\usage{
zIntervals(G)
}
\arguments{
\item{G}{A symmetric matrix object.}
}
\value{
This function gives a confident interval for Bernoulli distribution
}
\description{
Z test for Generalized Bernoulli Models
}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge university press.
}
\author{
Alejandro Espinosa-Rada
}
