% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zone_sample}
\alias{zone_sample}
\title{Zone-2 sampling from second-mode}
\usage{
zone_sample(A, X, ego = TRUE, core = FALSE)
}
\arguments{
\item{A}{A symmetric matrix object.}

\item{X}{X an incidence matrix object.}

\item{ego}{Whether to add or not ego into the subgraph.}

\item{core}{Whether to add actors at distance one from ego}
}
\value{
This function return a list of second-zone subgraphs using as a focal actor the second-mode of the multilevel network.
Each subgraph is the binary adjacency matrix of the meta-matrix of \code{A} and \code{X} restricted to the
nodes of the zone, without loops. When \code{core = TRUE}, each matrix has an attribute \code{core}, a named
vector that is one for the actors at distance one from the focal node and zero otherwise.
}
\description{
Second-zone multilevel sampling considering a second-mode focal actor
}
\examples{

A <- matrix(c(
  0, 1, 0, 0, 0, 0, 0, 0,
  0, 0, 1, 0, 0, 0, 0, 0,
  0, 1, 0, 1, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 1, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0
), byrow = TRUE, ncol = 8)
colnames(A) <- c("1", "2", "3", "4", "5", "6", "7", "8")
rownames(A) <- c("1", "2", "3", "4", "5", "6", "7", "8")

X <- matrix(c(
  1, 0, 0, 0,
  1, 0, 0, 0,
  1, 0, 1, 0,
  0, 1, 1, 0,
  0, 1, 1, 1,
  0, 1, 0, 0,
  0, 0, 0, 0,
  0, 0, 0, 1
), byrow = TRUE, ncol = 4)
colnames(X) <- c("a", "b", "c", "d")
rownames(X) <- c("1", "2", "3", "4", "5", "6", "7", "8")

set.seed(18051889)
zone_sample(A, X, core = TRUE)
}
\references{
Espinosa-Rada, A. (2021). A Network Approach for the Sociological Study of Science: Modelling Dynamic Multilevel Networks. [PhD](https://research.manchester.ac.uk/en/studentTheses/a-network-approach-for-the-sociological-study-of-science-and-know). The University of Manchester.
}
\author{
Alejandro Espinosa-Rada
}
