% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{components_id}
\alias{components_id}
\title{Components}
\usage{
components_id(A)
}
\arguments{
\item{A}{A matrix}
}
\value{
A vector assigning an id the components that each of the nodes of the matrix belongs
}
\description{
This function assigns an id to the components that each of the nodes of the matrix belongs
}
\examples{

A <- matrix(c(
  0, 1, 1, 0, 0,
  1, 0, 1, 0, 0,
  1, 1, 0, 0, 0,
  0, 0, 0, 0, 1,
  0, 0, 0, 1, 0
), byrow = TRUE, ncol = 5)
rownames(A) <- letters[1:ncol(A)]
colnames(A) <- rownames(A)
components_id(A)
}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
