% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{dist_sim_matrix}
\alias{dist_sim_matrix}
\title{Structural similarities}
\usage{
dist_sim_matrix(A, method = c("euclidean", "jaccard"))
}
\arguments{
\item{A}{A matrix}

\item{method}{The similarities/distance currently available are either \code{Euclidean} (default) or \code{Jaccard}.}
}
\value{
This function returns a distance matrix between nodes of the same matrix.
}
\description{
Structural similarities
}
\examples{
A <- matrix(c(
  0, 1, 0, 0, 1,
  0, 0, 0, 1, 1,
  0, 1, 0, 0, 1,
  0, 0, 1, 1, 0,
  0, 1, 0, 0, 0
), nrow = 5, ncol = 5, byrow = TRUE)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- rownames(A)
dist_sim_matrix(A, method = "jaccard")

A <- matrix(c(
  0, 0, 3, 0, 5,
  0, 0, 2, 0, 4,
  5, 4, 0, 4, 0,
  0, 3, 0, 1, 0,
  0, 0, 0, 0, 2
), nrow = 5, ncol = 5, byrow = TRUE)
dist_sim_matrix(A, method = "euclidean")
}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
