% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohesive_subgroups.R
\name{dyad_triad_table}
\alias{dyad_triad_table}
\title{Forbidden triad table}
\usage{
dyad_triad_table(A, adjacency_list = FALSE, min = NULL, max = NULL)
}
\arguments{
\item{A}{A symmetric matrix object.}

\item{adjacency_list}{Whether to return the adjacency list of triads 201 per node.}

\item{min}{Numeric constant, lower limit on the size of the triads 201 to find. NULL means no limit, ie. it is the same as 0.}

\item{max}{Numeric constant, upper limit on the size of the triads 201 to find. NULL means no limit.}
}
\value{
This function return the list of triads that each node belong.

If \code{adjacency_list = TRUE} it also  return the adjacency list of
the 'forbidden triads' per node.
}
\description{
This function explores dyads and triads (Simmel, 1950), building from the
'forbidden triad' (Granovetter, 1973).
First, the minimum structure is an isolated node, then dyads.
Afterwards, different combinations of 'forbidden triads' are explored.
}
\examples{
A <- matrix(c(
  0, 1, 1, 1, 0,
  1, 0, 1, 0, 0,
  1, 1, 0, 0, 0,
  1, 0, 0, 0, 1,
  0, 0, 0, 1, 0
), byrow = TRUE, ncol = 5)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- letters[1:ncol(A)]

dyad_triad_table(A, adjacency_list = TRUE, min = 3)
}
\references{
Granovetter, M.S. (1973). The Strength of Weak Ties. American Journal of Sociology. 78 (6): 1360–80. https://doi.org/10.1086/225469.

Simmel, G. (1950). Individual and Society. In K. H. Wolff (Ed.), The Sociology of George Simmel. New York: Free Press.

Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
