% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{dyadic_census}
\alias{dyadic_census}
\title{Dyadic census}
\usage{
dyadic_census(G, directed = TRUE)
}
\arguments{
\item{G}{A symmetric matrix object.}

\item{directed}{Whether the matrix is directed or not}
}
\value{
This function return the counts of the dyadic census.
}
\description{
Dyadic census
}
\examples{

data(krackhardt_friends)
dyadic_census(krackhardt_friends)

data(FIFAin)
dyadic_census(FIFAin[[1]], directed=FALSE)

}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
