% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_holes.R
\name{eb_constraint}
\alias{eb_constraint}
\title{Constraint}
\usage{
eb_constraint(A, ego = NULL, digraph = FALSE, weighted = FALSE)
}
\arguments{
\item{A}{A symmetric matrix object}

\item{ego}{Name of ego in the matrix}

\item{digraph}{Whether the  matrix is directed or undirected}

\item{weighted}{Whether the matrix is weighted or not}
}
\value{
This function returns term 1, 2 and 3, the normalization and the maximum value of the specification of Everett and Borgatti (2020),
and the constraint of Burt (1992).
}
\description{
Everett and Borgatti specification of the constraint measure for binary matrixes
}
\examples{

A <- matrix(c(0,1,1,0,0,1,
             1,0,1,0,0,1,
             1,1,0,0,0,1,
             0,0,0,0,1,1,
             0,0,0,1,0,1,
             1,1,1,1,1,0), ncol=6, byrow=TRUE)

rownames(A) <- letters[1:nrow(A)]
colnames(A) <- letters[1:ncol(A)]
eb_constraint(A, ego='f')

}
\references{
Burt, R.S., 1992. Structural Holes: the Social Structure of Competition. Harvard University Press, Cambridge.

Everett, M.G. and Borgatti, S., 2020. Unpacking Burt's constraint measure. Social Networks 62, pp. 50-57. doi: \url{https://doi.org/10.1016/j.socnet.2020.02.001}
}
\author{
Alejandro Espinosa-Rada
}
