% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{edgelist_to_matrix}
\alias{edgelist_to_matrix}
\title{Transform an edgelist to a matrix}
\usage{
edgelist_to_matrix(
  E,
  digraph = TRUE,
  label = NULL,
  label2 = NULL,
  bipartite = FALSE
)
}
\arguments{
\item{E}{An edge list}

\item{digraph}{Whether the matrix is directed or not}

\item{label}{A vector with the names of the nodes}

\item{label2}{A vector with the names of a different set of nodes}

\item{bipartite}{Whether the matrix is bipartite}
}
\value{
This function transform the edgelist into a matrix
}
\description{
Transform an edgelist to a matrix
}
\examples{
A <- matrix(c(
  0, 1, 1, 0, 0, 0, 0, 1, 0,
  1, 0, 1, 0, 0, 0, 0, 0, 0,
  1, 1, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 1, 1, 0, 0, 0,
  0, 0, 0, 1, 0, 0, 0, 0, 0,
  0, 0, 0, 1, 0, 0, 1, 1, 0,
  0, 0, 0, 0, 0, 1, 0, 1, 0,
  1, 0, 0, 0, 0, 1, 1, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0
), byrow = TRUE, ncol = 9)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- rownames(A)
E <- matrix_to_edgelist(A)
edgelist_to_matrix(E, label = c("i"), digraph = FALSE)
}
\author{
Alejandro Espinosa-Rada
}
