% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ego_net}
\alias{ego_net}
\title{Ego network}
\usage{
ego_net(
  A,
  ego = NULL,
  bipartite = FALSE,
  addEgo = FALSE,
  select = c("all", "in", "out")
)
}
\arguments{
\item{A}{A symmetric matrix object}

\item{ego}{Name of ego in the matrix}

\item{bipartite}{Whether the matrix is a two-mode network}

\item{addEgo}{Whether to retain ego in the submatrix or not}

\item{select}{Whether to consider all sender and receiver ties of ego (\code{all}), only incoming ties (\code{in}), or outgoing ties (\code{out}). By default, \code{all}.}
}
\value{
This function returns redundancy, effective size and efficincy measures (Burt, 1992).
}
\description{
Submatrix of ego's neighbourhoods
}
\examples{

A <- matrix(c(
  0, 1, 0, 0, 1, 1, 1,
  1, 0, 0, 1, 0, 0, 1,
  0, 0, 0, 0, 0, 0, 1,
  0, 1, 0, 0, 0, 0, 1,
  1, 0, 0, 0, 0, 0, 1,
  1, 0, 0, 0, 0, 0, 1,
  1, 1, 1, 1, 1, 1, 0
), ncol = 7, byrow = TRUE)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- letters[1:ncol(A)]
ego_net(A, ego = "g")
}
\references{
Burt, R.S., 1992. Structural Holes: the Social Structure of Competition. Harvard University Press, Cambridge.

Borgatti, S., 1997. Unpacking Burt's redundancy measure. Connections, 20(1): 35-38. doi: \url{http://www.analytictech.com/connections/v20(1)/holes.htm}
}
\author{
Alejandro Espinosa-Rada
}
