% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homophily.R
\name{ei_index}
\alias{ei_index}
\title{Krackhardt and Stern's E-I index}
\usage{
ei_index(A, mixed = TRUE, att = NULL)
}
\arguments{
\item{A}{A symmetric matrix object}

\item{mixed}{Whether the matrix provided is already a mixed matrix or not}

\item{att}{Categorical attribute of the nodes}
}
\value{
Numerical value of the E-I index.
}
\description{
This index was proposed by Krackhardt and Stern (1988) to distinguish between the relative prevalence
of between and within-group ties. This measure can be interpreted as homophily at the network level.
}
\examples{

n <- 100
A <- matrix(c(rbinom(n, 1, 0.5)),
  ncol = sqrt(n), nrow = sqrt(n), byrow = TRUE
)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- letters[1:ncol(A)]
att <- rbinom(sqrt(n), 3, 0.5)
ei_index(A, mixed = FALSE, att = att)
}
