% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_matrix}
\alias{expand_matrix}
\title{Expand Matrix}
\usage{
expand_matrix(A, label = NULL, loops = FALSE)
}
\arguments{
\item{A}{A square matrix}

\item{label}{Duplicated labels to expand the matrix}

\item{loops}{Whether the loops are retained or not}
}
\value{
Return an expanded matrix
}
\description{
Expand Matrix
}
\examples{
A <- matrix(c(
  0, 1, 1,
  0, 0, 1,
  1, 0, 0
), byrow = TRUE, ncol = 3, nrow = 3)
rownames(A) <- letters[1:NROW(A)]
colnames(A) <- rownames(A)
label <- sort(rep(rownames(A), 2))
expand_matrix(A, label, loops = FALSE)
}
\author{
Alejandro Espinosa-Rada
}
