% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{fractional_approach}
\alias{fractional_approach}
\title{Fractional approach}
\usage{
fractional_approach(
  A1,
  A2,
  approach = c("citation", "cocitation", "bcoupling")
)
}
\arguments{
\item{A1}{From incidence matrix (e.g. paper and authors)}

\item{A2}{To incidence matrix (e.g. author to paper)}

\item{approach}{Character string, \dQuote{citation}, \dQuote{cocitation} and \dQuote{bcoupling}}
}
\value{
Return a type of "citation network"
}
\description{
Matrix transformation from incidence matrices to citation networks, fractional counting for co-citation or fractional counting for bibliographic coupling
}
\examples{

A1 <- matrix(c(
  1, 0, 0, 0,
  0, 1, 0, 0,
  0, 1, 1, 1,
  0, 0, 0, 0,
  0, 0, 0, 1
), byrow = TRUE, ncol = 4)

A2 <- matrix(c(
  1, 1, 1, 0, 0,
  0, 0, 1, 0, 0,
  0, 0, 1, 1, 0,
  0, 0, 0, 1, 1
), byrow = TRUE, ncol = 5)

fractional_approach(A1, A2)
}
\references{
Batagelj, V. (2020). Analysis of the Southern women network using fractional approach. Social Networks, 68, 229-236 \url{https://doi.org/10.1016/j.socnet.2021.08.001}

Batagelj, V., & Cerinšek, M. (2013). On bibliographic networks. Scientometrics, 96(3), 845–864. \url{https://doi.org/10.1007/s11192-012-0940-1}
}
\author{
Alejandro Espinosa-Rada
}
