% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gen_density}
\alias{gen_density}
\title{Generalized density}
\usage{
gen_density(A, directed = TRUE, bipartite = FALSE, loops = FALSE)
}
\arguments{
\item{A}{A symmetric or incident matrix object}

\item{directed}{Wheter the matrix is directed or not}

\item{bipartite}{Wheter the matrix is bipartite or not.}

\item{loops}{Wheter to consider or not the loops}
}
\value{
This function returns the density of the matrix.
}
\description{
Generalized density
}
\examples{

B <- matrix(c(1,1,0,
              0,0,1,
              0,1,1,
              0,0,1), byrow=TRUE, ncol=3)
gen_density(B, bipartite = TRUE)


}
\references{
Wasserman, S., and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
