% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homophily.R
\name{heterogeneity}
\alias{heterogeneity}
\title{Blau's and IQV Index}
\usage{
heterogeneity(att, normalized = FALSE)
}
\arguments{
\item{att}{Categorical attribute of the nodes}

\item{normalized}{Whether to return IQV index}
}
\value{
Numerical value of the Blau index.

If \code{normalized = TRUE}, then the function also return IQV index.
}
\description{
This index was proposed by Blau (1977) to distinguish between the relative prevalence 
of between and within-group ties. This measure can be interpreted as homophily at the network level.
}
\examples{

n <- 100
att <- rbinom(sqrt(n), 3, 0.5)
heterogeneity(n, normalized = TRUE)

}
\references{
Agresti, A. and Agresti, B. (1978). Statistical Analysis of Qualitative Variation. Sociological Methodology, 9, 204-237. doi: \url{https://doi.org/10.2307/270810}   

Blau, P. M. (1977). Inequality and heterogeneity. New York: Free Press.
}
