% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad_uman.R
\name{hollandLeinhardt}
\alias{hollandLeinhardt}
\title{Triad census analysis assuming U|MAN}
\usage{
hollandLeinhardt(A, ztest = FALSE, covar = FALSE)
}
\arguments{
\item{A}{A symmetric matrix object.}

\item{ztest}{Whether to perform a z-test.}

\item{covar}{Whether to return the covarianc matrix for triadic analysis.}
}
\value{
This function gives the counts of the triad census, the expected counts, 
assuming that U|MAN distribution is operating, and the standard deviations of these counts.
}
\description{
Triad census analysis assuming U|MAN
}
\examples{

# COVARIANCE ARE WRONG IN W-F 1994? 

library(NetData)
data(kracknets, package = "NetData")
friend <- subset(friendship_data_frame, friendship_tie==1)

library(igraph)
g1 <- graph.data.frame(friend)
krackhardt <- get.adjacency(g1, sparse = FALSE)

triadCensusExp <- hollandLeinhardt(krackhardt)
triadCensusExp

\dontrun{
triadCensusExp <- hollandLeinhardt(krackhardt, ztest=TRUE, covar=TRUE)
triadCensusExp
}

}
\references{
Holland, P. W. and Leinhardt, S. (1975). The statistical analysis of local structure in social networks. In D. R. Heise (Ed.), Sociological Methodology, 1976 (Jossey-Bass, pp. 1–45).
Holland, P. W. and Leinhardt, S. (1976). Local Structure in Social Networks. Sociological Methodology, 7, 1–45.  doi: \url{https://doi.org/10.2307/270703}
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications (Vol. 8). Cambridge university press.
}
