% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard similarity}
\usage{
jaccard(
  A,
  B,
  directed = TRUE,
  diag = FALSE,
  coparticipation = FALSE,
  bipartite = TRUE
)
}
\arguments{
\item{A}{Binary matrix A}

\item{B}{Binary matrix B}

\item{directed}{Whether the matrix is symmetric}

\item{diag}{Whether the diagonal should be considered}

\item{coparticipation}{Select nodes that co-participate in both matrices}

\item{bipartite}{Whether the matrix is incident}
}
\value{
The output are: \code{jaccard} = Jaccard similarity, \code{proportion} =
proportion among the ties present at a given observation of ties that
are also present in the other matrix, and \code{table} = a table with the
tie changes between matrices.

If \code{coparticipation = TRUE}, then
also: \code{match} = The number of nodes present in both matrices;
 \code{size_matrix1} = The size of the first matrix;
 \code{size_matrix2} = The size of the second matrix;
\code{coparticipation1} = The percentage of nodes in the first matrix also present in the second matrix;
\code{coparticipation2} = The percentage of nodes in the second matrix also present in the first matrix:
\code{overlap_actors} = Overlap of nodes between two matrices

#' If \code{coparticipation = TRUE} and \code{bipartite = TRUE}, then
also: \code{matchM1} = The number of nodes in the first 'mode' present in both matrices;
\code{matchM2} = The number of nodes in the second 'mode' present in both matrices;
 \code{size_matrix1_M1} = The number of nodes in the first 'mode' of the first matrix;
 \code{size_matrix1_M2} = The number of nodes in the second 'mode' of the first matrix;
 \code{size_matrix2_M1} = The number of nodes in the first 'mode' of the second matrix;
 \code{size_matrix2_M2} = The number of nodes in the second 'mode' of the second matrix;
\code{coparticipation1_M2} = The percentage of nodes of the first 'mode' in the first matrix present in the second matrix.
\code{coparticipation1_M2} = The percentage of nodes of the second 'mode' in the first matrix present in the second matrix.
\code{coparticipation2_M1} = The percentage of nodes of the first 'mode' in the second matrix present in the first matrix.
 \code{coparticipation2_M2} = The percentage of nodes of the second 'mode' in the second matrix present in the first matrix.
\code{overlap_actors_M1} = Overlap between two matrices (nodes of the first 'mode')
\code{overlap_actors_M2} = Overlap between two matrices (nodes of the second 'mode')
}
\description{
Jaccard similarity identifies the changes of ties between two matrices.
}
\examples{

A <- matrix(c(
  0, 1, 1, 0,
  1, 0, 0, 0,
  1, 0, 0, 0,
  0, 0, 1, 0
), byrow = TRUE, ncol = 4)
B <- matrix(c(
  0, 1, 1, 0,
  1, 0, 0, 0,
  1, 0, 0, 0,
  0, 0, 0, 0
), byrow = TRUE, ncol = 4)
jaccard(A, B, directed = TRUE)
}
\references{
Batagelj, V., and Bren, M. (1995). Comparing resemblance measures. Journal of Classification 12, 73–90.
}
\author{
Alejandro Espinosa-Rada
}
