% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard similarity}
\usage{
jaccard(A, B, directed = TRUE, diag = FALSE, coparticipation = FALSE)
}
\arguments{
\item{A}{Binary matrix A}

\item{B}{Binary matrix B}

\item{directed}{Whether the matrix is symmetric}

\item{diag}{Whether the diagonal should be considered}

\item{coparticipation}{Select nodes that co-participate in both matrices}
}
\value{
The output are: \code{jaccard} = Jaccard similarity, \code{proportion} = 
proportion among the ties present at a given observation of ties that 
are also present in the other matrix, and \code{table} = a table with the 
tie changes between matrices.

If \code{coparticipation = TRUE}, then 
also: \code{match} = number of nodes present in both matrices;
\code{coparticipation1} = percentage 
of actors in the first matrix also present in the 
second matrix; \code{coparticipation2}` = percentage 
of actors in the second matrix also present in the first matrix
}
\description{
Jaccard similarity to identify the tie changes between two matrices.
}
\examples{

A <- matrix(c(0,1,1,0,
              1,0,0,0,
              1,0,0,0,
              0,0,1,0), byrow=TRUE, ncol= 4)
B <- matrix(c(0,1,1,0,
              1,0,0,0,
              1,0,0,0,
              0,0,0,0), byrow=TRUE, ncol= 4)
jaccard(A, B, directed = TRUE)

}
\references{
Batagelj, V., and Bren, M. (1995). Comparing resemblance measures. Journal of Classification 12, 73–90.
}
\author{
Alejandro Espinosa-Rada
}
