% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{k_core}
\alias{k_core}
\title{Generalized K-core}
\usage{
k_core(
  A,
  B1 = NULL,
  multilevel = FALSE,
  type = "in",
  digraph = FALSE,
  loops = FALSE,
  weighted = FALSE,
  alpha = 1
)
}
\arguments{
\item{A}{A symmetric matrix object.}

\item{B1}{An incident matrix for multilevel networks.}

\item{multilevel}{Wheter the measure of k-core is for multilevel networks.}

\item{type}{Character string, \dQuote{out} (outdegree), \dQuote{in} (indegree) and \dQuote{all} (degree)}

\item{digraph}{Wheter the  matrix is directed or undirected}

\item{loops}{Wheter the diagonal of the matrix is considered or not}

\item{weighted}{Wheter the measure of k-core is for valued matrices}

\item{alpha}{Sets the alpha parameter in the generalised measures from Opsahl et al. (2010)}
}
\value{
This function return the k-core.
}
\description{
Generalized k-core for undirected, directed, weighted and multilevel networks
}
\examples{

A1 <- matrix(c(0,1,0,0,0,
               1,0,0,1,0,
               0,0,0,1,0,
               0,1,1,0,1,
               0,0,0,1,0), byrow=TRUE, ncol=5)
B1 <- matrix(c(1,0,0,
               1,1,0,
               0,1,0,
               0,1,0,
               0,1,1), byrow=TRUE, ncol=3)

k_core(A1, B1, multilevel=TRUE)

}
\references{
Batagelj, V., & Zaveršnik, M. (2011). Fast algorithms for determining (generalized) core groups in social networks. Advances in Data Analysis and Classification, 5(2), 129–145. https://doi.org/10.1007/s11634-010-0079-y

Eidsaa, M., & Almaas, E. (2013). $s$-core network decomposition: A generalization of $k$-core analysis to weighted networks. Physical Review E, 88(6), 062819. https://doi.org/10.1103/PhysRevE.88.062819

Espinosa-Rada, Alejandro (forthcoming) PhD Thesis. 

 Seidman S (1983).  'Network structure and minimum degree'.  Social Networks, 5, 269-287.
}
\author{
Alejandro Espinosa-Rada
}
