% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{mixed_census}
\alias{mixed_census}
\title{Multilevel triad and quadrilateral census}
\usage{
mixed_census(A1, B1, B2 = NULL, quad = FALSE)
}
\arguments{
\item{A1}{An adjacent matrix object.}

\item{B1}{An incidence matrix object.}

\item{B2}{An incidence matrix object.}

\item{quad}{Whether the matrix is a quadrilateral census or not.}
}
\value{
This function return the counts of a multilevel census.

If \code{quad = TRUE}, then the function return the multilevel quadrilateral census.
}
\description{
Multilevel triad and quadrilateral census
}
\examples{

B1 <- matrix(c(
  1, 1, 0,
  0, 0, 1,
  0, 0, 1,
  1, 0, 0
), byrow = TRUE, ncol = 3)
A1 <- matrix(c(
  0, 1, 0, 1,
  1, 0, 0, 1,
  0, 1, 0, 1,
  1, 0, 1, 0
), byrow = TRUE, ncol = 4)
B2 <- matrix(c(
  1, 0, 0, 0, 0,
  0, 1, 0, 1, 0,
  0, 0, 0, 0, 0,
  0, 0, 0, 0, 0
), byrow = TRUE, ncol = 5)

mixed_census(A1, B1, B2, quad = TRUE)
}
\references{
Espinosa-Rada, A. (2021). A Network Approach for the Sociological Study of Science: Modelling Dynamic Multilevel Networks. [PhD](https://research.manchester.ac.uk/en/studentTheses/a-network-approach-for-the-sociological-study-of-science-and-know). The University of Manchester.

Espinosa-Rada, A., Bellotti, E., Everett, M., & Stadtfeld, C. (2024). Co-evolution of a socio-cognitive scientific network: A case study of citation dynamics among astronomers. Social Networks, 78, 92–108. https://doi.org/10.1016/j.socnet.2023.11.008

Hollway, J., Lomi, A., Pallotti, F., & Stadtfeld, C. (2017). Multilevel social spaces: The network dynamics of organizational fields. Network Science, 5(2), 187–212. https://doi.org/10.1017/nws.2017.8
}
\author{
Alejandro Espinosa-Rada
}
