% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{multilevel_degree}
\alias{multilevel_degree}
\title{Degree centrality for multilevel networks.}
\usage{
multilevel_degree(
  A1,
  B1,
  A2 = NULL,
  B2 = NULL,
  A3 = NULL,
  B3 = NULL,
  complete = FALSE,
  digraphA1 = FALSE,
  digraphA2 = FALSE,
  digraphA3 = FALSE,
  typeA1 = "out",
  typeA2 = "out",
  typeA3 = "out",
  loopsA1 = FALSE,
  loopsA2 = FALSE,
  loopsA3 = FALSE,
  normalized = FALSE,
  weightedA1 = FALSE,
  weightedA2 = FALSE,
  weightedA3 = FALSE,
  alphaA1 = 0.5,
  alphaA2 = 0.5,
  alphaA3 = 0.5
)
}
\arguments{
\item{A1}{The adjacent square matrix of the lowest level}

\item{B1}{The incident matrix of the ties between the nodes of first level and the nodes of the second level}

\item{A2}{The adjacent square of the second level}

\item{B2}{The incident matrix of the ties between the nodes of the second level and the nodes of the third level}

\item{A3}{The adjacent square of the third level}

\item{B3}{The incident matrix of the ties between the nodes of the third level and the nodes of the first level}

\item{complete}{Add the degree of bipartite and tripartite networks for B1, B2 and/or B3, and the low_multilevel (i.e. A1+B1+B2+B3), meso_multilevel (i.e. B1+A2+B2+B3) and high_multilevel (i.e. B1+B2+A3+B3) degree}

\item{digraphA1}{Wheter A1 is a directed network}

\item{digraphA2}{Wheter A2 is a directed network}

\item{digraphA3}{Wheter A3 is a directed network}

\item{typeA1}{Type of degree of the network for A1, "out" for out-degree, "in" for in-degree or "all" for the sum of the two}

\item{typeA2}{Type of degree of the network for A2, "out" for out-degree, "in" for in-degree or "all" for the sum of the two}

\item{typeA3}{Type of degree of the network for A3, "out" for out-degree, "in" for in-degree or "all" for the sum of the two}

\item{loopsA1}{Wheter the loops of the edges are considered in matrix A1}

\item{loopsA2}{Wheter the loops of the edges are considered in matrix A2}

\item{loopsA3}{Wheter the loops of the edges are considered in matrix A3}

\item{normalized}{If TRUE then the result is divided by (n-1)+k+m for the first level, (m-1)+n+k for the second level, and (k-1)+m+n according to Espinosa-Rada et al. (2021)}

\item{weightedA1}{Wheter A1 is weighted}

\item{weightedA2}{Wheter A2 is weighted}

\item{weightedA3}{Wheter A3 is weighted}

\item{alphaA1}{The alpha parameter of A1 according to Opsahl et al (2010) for weighted networks. The value 0.5 is given by default.}

\item{alphaA2}{The alpha parameter of A2 according to Opsahl et al (2010) for weighted networks. The value 0.5 is given by default.}

\item{alphaA3}{The alpha parameter of A3 according to Opsahl et al (2010) for weighted networks. The value 0.5 is given by default.}
}
\value{
Return a data.frame of multilevel degree
}
\description{
Degree centrality for multilevel networks.
}
\examples{

A1 <- matrix(c(0,1,0,0,0,
              1,0,0,1,0,
              0,0,0,1,0,
              0,1,1,0,1,
              0,0,0,1,0), byrow=TRUE, ncol=5)
              
B1 <- matrix(c(1,0,0,
              1,1,0,
              0,1,0,
              0,1,0,
              0,1,1), byrow=TRUE, ncol=3)

A2 <- matrix(c(0,1,1,
              1,0,0,
              1,0,0), byrow=TRUE, nrow=3)

B2 <- matrix(c(1,1,0,0,
              0,0,1,0,
              0,0,1,1), byrow=TRUE, ncol=4)

A3 <- matrix(c(0,1,1,1,
              1,0,0,0,
              1,0,0,1,
              1,0,1,0), byrow=TRUE, ncol=4)

B3 <- matrix(c(1,0,0,0,0,
              0,1,0,1,0,
              0,0,0,0,0,
              0,0,0,0,0), byrow=TRUE, ncol=5)

multilevel_degree(A1, B1, A2, B2, A3, B3)

\dontrun{
multilevel_degree(A1, B1, A2, B2, A3, B3, normalized=TRUE, complete=TRUE)
}

}
\references{
Borgatti, S. P., and Everett, M. G. (1997). Network analysis of 2-mode data. Social Networks, 19(3), 243–269.

Espinosa-Rada, Alejandro; Bellotti, Elisa; Everett, Martin & Stadtfeld, Christoph (forthcoming). "Co-evolution of Scientific Networks: Multilevel Analysis of a National Discipline"

Freeman, L. C. (1978). Centrality in social networks conceptual clarification. Social Networks, 1(3), 215–239.

Opsahl, T., Agneessens, F., and Skvoretz, J. (2010). Node centrality in weighted networks: Generalizing degree and shortest paths. Social Networks, 32(3), 245–251.
}
\author{
Alejandro Espinosa-Rada
}
