% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simplicial_complexes}
\alias{simplicial_complexes}
\title{Simplicial Complexes}
\usage{
simplicial_complexes(A, zero_simplex = FALSE, projection = FALSE)
}
\arguments{
\item{A}{A symmetric matrix object.}

\item{zero_simplex}{Whether to include the zero simple.}

\item{projection}{Whether to return the links between actors (i.e., rows) through their shared linking events (i.e., columns).}
}
\value{
This function return an incidence matrix of actors participating in simplices or simplicial complexes
}
\description{
incidence matrix of simplexes or cliques
}
\examples{
A <- matrix(c(
  0, 1, 1, 0, 0, 0, 0, 1, 0,
  1, 0, 1, 0, 0, 0, 0, 0, 0,
  1, 1, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 1, 1, 0, 0, 0,
  0, 0, 0, 1, 0, 0, 0, 0, 0,
  0, 0, 0, 1, 0, 0, 1, 1, 0,
  0, 0, 0, 0, 0, 1, 0, 1, 0,
  1, 0, 0, 0, 0, 1, 1, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0
), byrow = TRUE, ncol = 9)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- rownames(A)
simplicial_complexes(A, zero_simplex = FALSE)
}
\references{
Atkin, R. H. (1974). Mathematical structure in human affairs. New York: Crane, Rusak.

Freeman, L. C. (1980). Q-analysis and the structure of friendship networks. International Journal of Man-Machine Studies, 12(4), 367–378. https://doi.org/10.1016/S0020-7373(80)80021-6

Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
