% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed.R
\name{struc_balance}
\alias{struc_balance}
\title{Structural balance}
\usage{
struc_balance(A)
}
\arguments{
\item{A}{A signed matrix (i.e., with ties that are either -1, 0 or 1)}
}
\value{
This function return the structural balance (Heider, 1940; Cartwright and Harary, 1956)
}
\description{
Structural balance
}
\examples{
B <- matrix(c(0,-1,-1,0,
              -1,0,1,0,
              -1,1,0,0,
              0,0,0,0), byrow = TRUE, ncol = 4)
rownames(B) <- letters[1:nrow(B)]
colnames(B) <- rownames(B)
struc_balance(B)

}
\references{
Cartwright, Dorwin, and Harary, Frank (1956). Structural balance: a generalization of Heider's theory. Psychological review, 63(5), 277.

Heider, Fritz (1946). "Attitudes and Cognitive Organization". The Journal of Psychology. 21: 107–112
}
\author{
Alejandro Espinosa-Rada
}
