% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{structural_na}
\alias{structural_na}
\title{Structural missing data}
\usage{
structural_na(A, label = NULL, bipartite = FALSE, column = FALSE)
}
\arguments{
\item{A}{A symmetric or incident matrix object}

\item{label}{String vector with the names of the theoretical complete matrix}

\item{bipartite}{Whether the matrix is bipartite or not.}

\item{column}{Whether the assignation of NA is for columns in the biparite network, row by default.}
}
\value{
This function returns NA to missing data.
}
\description{
Assign NA to missing data in the matrices
}
\examples{

A <- matrix(c(
  0, 1, 1,
  1, 0, 1,
  0, 0, 0
), byrow = TRUE, ncol = 3)
colnames(A) <- c("A", "C", "D")
rownames(A) <- c("A", "C", "D")
label <- c("A", "B", "C", "D", "E")
structural_na(A, label)
}
\author{
Alejandro Espinosa-Rada
}
