% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad_uman.R
\name{triad_uman}
\alias{triad_uman}
\title{Triad census analysis assuming U|MAN}
\usage{
triad_uman(A, ztest = FALSE, covar = FALSE)
}
\arguments{
\item{A}{A symmetric matrix object.}

\item{ztest}{Return Z and p-value}

\item{covar}{Return the covarianc matrix for triadic analysis (EXPERIMENTAL).}
}
\value{
This function gives the counts of the triad census, the expected counts, 
assuming that U|MAN distribution is operating, and the standard deviations of these counts.
}
\description{
Triad census analysis assuming U|MAN
}
\examples{

data(krackhardt_friends)
triad_uman(krackhardt_friends)

\dontrun{
triad_uman(A, ztest=TRUE)
}

}
\references{
Holland, P. W. and Leinhardt, S. (1975). The statistical analysis of local structure in social networks. In D. R. Heise (Ed.), Sociological Methodology, 1976 (Jossey-Bass, pp. 1–45).

Holland, P. W. and Leinhardt, S. (1976). Local Structure in Social Networks. Sociological Methodology, 7, 1–45.  doi: \url{https://doi.org/10.2307/270703}

Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
