% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad_uman.R
\name{triad_uman}
\alias{triad_uman}
\title{Triad census analysis assuming U|MAN}
\usage{
triad_uman(A, covar = FALSE)
}
\arguments{
\item{A}{A symmetric matrix object.}

\item{covar}{Whether to return the covarianc matrix for triadic analysis.}
}
\value{
This function gives the counts of the triad census, the expected counts, 
assuming that U|MAN distribution is operating, and the standard deviations of these counts.
}
\description{
Triad census analysis assuming U|MAN
}
\examples{

# COVARIANCE ARE WRONG IN W-F 1994? 
# (i.e. TRIADS computer program of Walker and Wasserman, 1987)
# 201-102: we use the equation of Leinhardt and Holland (1975)
# P0(u,v): 9m^(3)n^(3); P1(u,v): 2m^(2)m^(2)(m+4-n)
# results: W-F: -18.4; here: 7.06
# 300-030T: original: -0.01; here:0.011

A <- matrix(c(0,1,0,1,0,0,1,0,0,1,0,0,0,1,0,0,0,0,0,0,0,
             1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,
             0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,1,0,0,0,0,
             1,1,0,0,0,0,1,0,0,1,0,0,0,1,1,0,0,0,0,0,0,
             0,1,0,0,0,0,0,0,1,0,0,1,0,0,1,0,1,0,1,1,0,
             0,1,0,0,0,0,0,0,0,1,0,0,0,0,1,0,0,0,1,1,1,
             0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
             0,0,1,0,1,0,1,0,0,1,0,0,0,1,0,0,0,1,0,1,0,
             1,1,1,1,1,0,1,0,0,1,1,0,1,0,1,1,1,0,0,1,0,
             1,0,0,1,0,0,0,0,0,0,0,0,0,0,1,0,0,0,1,0,0,
             0,0,0,0,1,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,
             0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,1,
             1,0,1,0,1,1,0,0,1,0,0,1,0,0,0,0,1,0,0,1,0,
             1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
             1,1,1,1,1,1,1,1,1,1,0,1,1,1,0,0,1,1,1,1,1,
             0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
             1,1,1,0,1,0,0,0,1,1,0,1,1,0,0,0,0,1,0,0,0,
             0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,1,0,0,0,0,0,
             0,1,0,0,0,0,0,0,0,1,0,0,0,0,1,1,0,0,0,0,0,
             0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
             0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0), 
             ncol=21, byrow=TRUE)

triadCensusExp <- triad_uman(A)
triadCensusExp

\dontrun{
triadCensusExp <- triad_uman(A, covar=TRUE)
triadCensusExp
}

}
\references{
Holland, P. W. and Leinhardt, S. (1975). The statistical analysis of local structure in social networks. In D. R. Heise (Ed.), Sociological Methodology, 1976 (Jossey-Bass, pp. 1–45).

Holland, P. W. and Leinhardt, S. (1976). Local Structure in Social Networks. Sociological Methodology, 7, 1–45.  doi: \url{https://doi.org/10.2307/270703}

Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge university press.
}
\author{
Alejandro Espinosa-Rada
}
