% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{zarc}
\alias{zarc}
\title{Z test for Arcs in Uniform Distribution}
\usage{
zarc(G)
}
\arguments{
\item{G}{A symmetric matrix object.}
}
\value{
This function gives a Z test and p-value for arcs in uniform distribution
}
\description{
Z test for Arcs in Uniform Distribution
}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge university press.
}
\author{
Alejandro Espinosa-Rada
}
