% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{zarc}
\alias{zarc}
\title{Z test of the number of arcs}
\usage{
zarc(G, p = 0.5, interval = FALSE)
}
\arguments{
\item{G}{A symmetric matrix object.}

\item{p}{Constant probability p.}

\item{interval}{Return a 95 percent confidence interval.}
}
\value{
This function gives a Z test and p-value for the number of lines or arcs present in a directed graph
}
\description{
Z test of the number of arcs
}
\examples{

data(krackhardt_friends)
zarc(krackhardt_friends)

}
\references{
Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.
}
\author{
Alejandro Espinosa-Rada
}
